split into separate files, add XML doc comments
